pyproject.toml
[tool.poetry]
name = "modapi"
version = "0.1.8"
description = "Unified API for Modbus communication"
authors = ["Tom Sapletta <info@softreck.dev>"]
readme = "README.md"
packages = [{include = "modapi"}]

[tool.poetry.scripts]
modapi = "modapi.__main__:main"
modbus = "modapi.cli:main"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
pymodbus = {extras = ["serial"], version = "^3.10.0"}
python-dotenv = "^0.19.0"
flask = "^2.0.0"
paho-mqtt = "^2.0.0"
requests = "^2.32.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
pytest-cov = "^4.0.0"
pytest-mock = "^3.11.1"
pytest-asyncio = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
